Move metadata common (protobuf) types to restate_types
Summary:
This is a planning step to move all metadata common protobuf
type to restate_types. This way the data types can
be reused in other crates and .proto files
without a direct dependency on metadata-server crate
