+++ crates/metadata-server/src/grpc/mod.rs
@@ -15,30 +15,12 @@ tonic::include_proto!("restate.metadata_server_svc");
 pub const FILE_DESCRIPTOR_SET: &[u8] = tonic::include_file_descriptor_set!("metadata_server_svc");
 
 pub mod pb_conversions {
-    use crate::grpc::{
-        GetResponse, GetVersionResponse, PreconditionKind, Ulid, WriteRequest, WriteRequestKind,
-    };
-    use crate::{MetadataServerSummary, grpc};
-    use restate_core::metadata_store::{Precondition, VersionedValue};
     use restate_types::Version;
+    use restate_types::errors::ConversionError;
+    use restate_types::metadata::VersionedValue;
 
-    #[derive(Debug, thiserror::Error)]
-    pub enum ConversionError {
-        #[error("missing field '{0}'")]
-        MissingField(&'static str),
-        #[error("invalid data '{0}'")]
-        InvalidData(&'static str),
-    }
-
-    impl ConversionError {
-        pub fn missing_field(field: &'static str) -> Self {
-            ConversionError::MissingField(field)
-        }
-
-        pub fn invalid_data(field: &'static str) -> Self {
-            ConversionError::InvalidData(field)
-        }
-    }
+    use crate::grpc::{GetResponse, GetVersionResponse, Ulid, WriteRequest, WriteRequestKind};
+    use crate::{MetadataServerSummary, grpc};
 
     impl TryFrom<GetResponse> for Option<VersionedValue> {
         type Error = ConversionError;
@@ -52,82 +34,9 @@ pub mod pb_conversions {
         }
     }
 
-    impl TryFrom<grpc::VersionedValue> for VersionedValue {
-        type Error = ConversionError;
-
-        fn try_from(value: grpc::VersionedValue) -> Result<Self, Self::Error> {
-            let version = value
-                .version
-                .ok_or_else(|| ConversionError::missing_field("version"))?;
-            Ok(VersionedValue::new(version.into(), value.bytes))
-        }
-    }
-
     impl From<GetVersionResponse> for Option<Version> {
         fn from(value: GetVersionResponse) -> Self {
             value.version.map(Into::into)
-        }
-    }
-
-    impl From<VersionedValue> for grpc::VersionedValue {
-        fn from(value: VersionedValue) -> Self {
-            grpc::VersionedValue {
-                version: Some(value.version.into()),
-                bytes: value.value,
-            }
-        }
-    }
-
-    impl From<grpc::Version> for Version {
-        fn from(value: grpc::Version) -> Self {
-            Version::from(value.value)
-        }
-    }
-
-    impl From<Version> for grpc::Version {
-        fn from(value: Version) -> Self {
-            grpc::Version {
-                value: value.into(),
-            }
-        }
-    }
-
-    impl From<Precondition> for grpc::Precondition {
-        fn from(value: Precondition) -> Self {
-            match value {
-                Precondition::None => grpc::Precondition {
-                    kind: PreconditionKind::None.into(),
-                    version: None,
-                },
-                Precondition::DoesNotExist => grpc::Precondition {
-                    kind: PreconditionKind::DoesNotExist.into(),
-                    version: None,
-                },
-                Precondition::MatchesVersion(version) => grpc::Precondition {
-                    kind: PreconditionKind::MatchesVersion.into(),
-                    version: Some(version.into()),
-                },
-            }
-        }
-    }
-
-    impl TryFrom<grpc::Precondition> for Precondition {
-        type Error = ConversionError;
-
-        fn try_from(value: grpc::Precondition) -> Result<Self, Self::Error> {
-            match value.kind() {
-                PreconditionKind::Unknown => {
-                    Err(ConversionError::invalid_data("unknown precondition kind"))
-                }
-                PreconditionKind::None => Ok(Precondition::None),
-                PreconditionKind::DoesNotExist => Ok(Precondition::DoesNotExist),
-                PreconditionKind::MatchesVersion => Ok(Precondition::MatchesVersion(
-                    value
-                        .version
-                        .ok_or_else(|| ConversionError::missing_field("version"))?
-                        .into(),
-                )),
-            }
         }
     }
