+++ crates/metadata-server/src/raft/kv_memory_storage.rs
@@ -9,15 +9,15 @@
 // by the Apache License, Version 2.0.
 
 use crate::grpc::MetadataServerSnapshot;
-use crate::grpc::pb_conversions::ConversionError;
 use crate::{
     Callback, PreconditionViolation, ReadOnlyRequest, ReadOnlyRequestKind, RequestError,
     RequestKind, WriteRequest, grpc,
 };
 use bytestring::ByteString;
 use restate_core::MetadataWriter;
-use restate_core::metadata_store::{Precondition, VersionedValue};
 use restate_types::Version;
+use restate_types::errors::ConversionError;
+use restate_types::metadata::{Precondition, VersionedValue};
 use restate_types::metadata_store::keys::NODES_CONFIG_KEY;
 use restate_types::nodes_config::NodesConfiguration;
 use restate_types::storage::StorageCodec;
