+++ server/tests/raft_metadata_cluster.rs
@@ -13,7 +13,7 @@ use enumset::EnumSet;
 use googletest::prelude::err;
 use googletest::{IntoTestResult, assert_that, pat};
 use rand::seq::IndexedMutRandom;
-use restate_core::metadata_store::{Precondition, WriteError};
+use restate_core::metadata_store::WriteError;
 use restate_core::{TaskCenter, TaskKind, cancellation_watcher};
 use restate_local_cluster_runner::cluster::Cluster;
 use restate_local_cluster_runner::node::{BinarySource, HealthCheck, Node};
@@ -23,6 +23,7 @@ use restate_types::Versioned;
 use restate_types::config::{
     Configuration, MetadataClientKind, MetadataClientOptions, MetadataServerKind, RaftOptions,
 };
+use restate_types::metadata::Precondition;
 use std::num::NonZeroUsize;
 use std::time::{Duration, Instant};
 use tracing::info;
